% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na-tests.R
\name{na-tests}
\alias{na-tests}
\alias{equally_na}
\alias{equally_not_na}
\alias{equally_na_or_not_na}
\title{Test for joint missingness}
\usage{
equally_na(x, y, strict = FALSE)

equally_not_na(x, y, strict = FALSE)

equally_na_or_not_na(x, y, strict = FALSE)
}
\arguments{
\item{x, y}{Two vectors. They should have the same length unless either has
length 1.}

\item{strict}{Logical. If set to \code{TRUE}, the missingness check results of
each pair of values are checked for exact equality using \code{\link[=identical]{identical()}}.
This includes equal attributes such as names. Default is \code{FALSE}.}
}
\value{
Logical. Never \code{NA}. The length depends on \code{x} and \code{y}, as in \code{\link{&}},
\code{\link{==}}, etc.
}
\description{
For each pair of values, these functions run checks for missing
values.

By default:
\itemize{
\item \code{equally_na()} checks that both values are \code{NA} (missing).
\item \code{equally_not_na()} checks that both values are not \code{NA}.
\item \code{equally_na_or_not_na()} checks that either both values are \code{NA}, or both
are not \code{NA}.
}
}
