% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order2.R
\name{order2}
\alias{order2}
\title{Correctly handle \code{NA} in \code{order()}}
\usage{
order2(
  ...,
  na.last = NULL,
  decreasing = FALSE,
  method = c("auto", "shell", "radix")
)
}
\arguments{
\item{..., na.last, decreasing, method}{Passed on to \code{\link[base:order]{base::order()}}. Note
that the default for \code{na.last} is \code{TRUE} in \code{\link[=order]{order()}}, but \code{NULL} in
\code{order2()}.}
}
\value{
Integer or double. See \code{\link[base:order]{base::order()}} for details.
}
\description{
Use \code{order2()} instead of \code{\link[=order]{order()}} if your data may contain
missing values that shouldn't simply be assumed to be larger or smaller
than the known values.

By default, missing values lead to an error. Specify \code{na.last} as in
\code{\link[base:order]{base::order()}} to override this.
}
\details{
\code{order2()} is just a wrapper around \code{\link[base:order]{base::order()}}. There is no
difference apart from missing value handling.

For the full rationale behind this function, see TODO: REFER TO BIG ARTICLE
ONCE IT'S DONE; SECTION ON \code{order()}!
}
