% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which2.R
\name{which2}
\alias{which2}
\title{Correctly handle \code{NA} in \code{which()}}
\usage{
which2(
  x,
  arr.ind = FALSE,
  useNames = TRUE,
  na.treat = c("error", "warning", "true", "false")
)
}
\arguments{
\item{x, arr.ind, useNames}{Passed down to \code{\link[base:which]{base::which()}}.}

\item{na.treat}{String. One of these options:
\itemize{
\item \code{"error"} (the default) throws an error if any elements of \code{x} are \code{NA}.
\item \code{"warning"} alerts you to \code{NA}s but handles them like \code{"false"}.
\item \code{"true"} treats \code{NA}s as \code{TRUE}, so their indices are part of the output.
\item \code{"false"} treats \code{NA}s as \code{FALSE}, so their indices are not part of the
output. This works just like \code{\link[base:which]{base::which()}}.
}}
}
\value{
Integer vector. Its length is never greater than \code{length(x)}, and it
never contains \code{NA}. See \code{\link[base:which]{base::which()}} for details.
}
\description{
Use \code{which2()} instead of \code{\link[=which]{which()}} if your data may contain
missing values that shouldn't simply be ignored.

Specify \code{na.treat} to decide how to handle missing values. By default, they
will lead to an error.
}
\details{
\code{which2()} is just a wrapper around \code{\link[base:which]{base::which()}}. There is no
difference apart from missing value handling.

For the full rationale behind this function, see TODO: REFER TO BIG ARTICLE
ONCE IT'S DONE; SECTION ON \code{which()}!
}
